# feat(event pool): simplify realization of event pool