# Description
I have managed to improve the performance of the pool and improve its readability

# Benchmarks
## Old
```
BenchmarkEventPoolOneGoroutine-8        58460538                19.96 ns/op            0 B/op          0 allocs/op
BenchmarkEventPoolManyGoroutines-8           100          10.746.244 ns/op            4986 B/op         56 allocs/op
BenchmarkEventPoolSlowestPath-8              152           8.189.235 ns/op           27087 B/op       1034 allocs/op
```
## New
```
BenchmarkEventPoolOneGoroutine-8        640555099               26.96 ns/op            0 B/op          0 allocs/op
BenchmarkEventPoolManyGoroutines-8         12850           1.404.986 ns/op             240 B/op          9 allocs/op
BenchmarkEventPoolSlowestPath-8            21242            810.480 ns/op           24170 B/op       1002 allocs/op
```